Register Screen Styles
Adding basic styles to the page to get it look generally close to the design provided by graphics. There are still some issues to work through such as resizing the input verification color status bars to match the widths of the inputs, getting the background image to show up behind the title bar, and getting the register button to always be at the bottom of the page.